RBAC in server + alerts in web

+++ apps/server/src/controllers/metrics/delete-metric.ts
@@ -3,6 +3,11 @@ import express from 'express'
 
 export const deleteMetric = async (req:express.Request,res:express.Response) => {
     try{    
+         const userRole = req.session.user?.userRole;
+        if(userRole === undefined  || ((userRole === "VIEWER") || (userRole === "MEMBER"))){
+            res.status(403).json({success : true,message : "Not Authorised"})
+            return;
+        }
         const metricId = req.params.metricId;
         
         const deleteMetric = await prisma.metrics.delete({
